Schema of the proposed economic upgrade
Results of the 3 economic practice available for personal choice to guide the money for expected utility :+1: Consumption, 2: Savings, 3: Investment
The upgrade proposed to this classic Economy is the option for the practice of the money datevaluation as a 4th gesture. 
Economy 4G3W is that provided with 3G+Datevaluation = 4, living in Internet throught the WUW, make 3W.
So here the upgrade is for the practice of all 4 gestures (datevaluation or investment substitute included) acting on to the W3C (may be suppose one 1st protocole) for  Economy 4G3W.
The thing depends on W3C protocole between Authorig WUW's Team, Google-Alphabet and USA's Administration.